Fixed empty message error

+++ commands/say.js
@@ -5,6 +5,10 @@ module.exports = {
 	usage: 'say [message]',
 	disabled: false,
 	execute(client, message, args) {
+		if (!args[0]) {
+			message.react('❌');
+			return;
+		}
 		message.channel.send(args.join(' '));
 		message.delete();
 	},
